Fix double-click navigation
- Fixed navigation by double-clicking bug
- Parsed sent optimization parameters and added a comment as how to get them
- Changed "Model" syntax to "Simulation"
- Updated delete requests according to permissions
- Fixed an error with model details if user has never saved their simulation parameters before
- Added new statistics
- Added a delete button for users to allow them delete their simulations
- Fixed a bug where public simulations could result an exception since they would not have explicit permissions
- Limited regular user model creation number in last 24 hours

+++ Controllers/BasinsController.cs
@@ -109,6 +109,18 @@ namespace HydroFlowProject.Controllers
         [Route("deleteBasin")]
         public async Task<ActionResult<Basin>> DeleteBasin([FromBody] Basin basin)
         {
+            var basinHasModels = _context.BasinModels.ToList().Find(bm => bm.BasinId == basin.Id);
+            if (basinHasModels != null)
+            {
+                return StatusCode(StatusCodes.Status412PreconditionFailed, "This basin has simulations created in it!");
+            }
+
+            var basinPerm = _context.BasinPermissions.FirstOrDefault(bm => bm.BasinId.Equals(basin.Id));
+            if (basinPerm != null)
+            {
+                _context.BasinPermissions.Remove(basinPerm);
+            }
+
             _context.Basins.Remove(basin);
             int result = await _context.SaveChangesAsync();
             if (result > 0)
@@ -186,10 +198,11 @@ namespace HydroFlowProject.Controllers
                     });
                 }
             }
-
+            var totalCount = _context.Database.SqlQuery<int>($"select count(distinct dbo.Basin_Models.ModelId) from dbo.Basin_Models\ninner join dbo.User_Models on dbo.User_Models.ModelId = dbo.Basin_Models.ModelId");
+            
             Dictionary<string, object> resultMap = new()
             {
-                { "totalCount", modelIdList.Count },
+                { "totalCount", totalCount },
                 { "modelList", modelList }
             };
